% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.clonevol.R
\name{extract.tables.from.clonevol}
\alias{extract.tables.from.clonevol}
\title{\if{html}{\out{
<Add Title>
}}}
\usage{
\method{extract}{tables.from.clonevol}(y, model = 1, explicit_parents = list())
}
\description{
\if{html}{\out{
<Add Description>
}}
}
