% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonevol_functions.R
\name{extract_tables_from_clonevol}
\alias{extract_tables_from_clonevol}
\title{\if{html}{\out{
<Add Title>
}}}
\usage{
extract_tables_from_clonevol(y, model = 1, explicit_parents = list())
}
\description{
\if{html}{\out{
<Add Description>
}}
}
