% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_tables}
\alias{filter_tables}
\title{Filter Jellyfish tables by patient}
\usage{
filter_tables(tables, patient)
}
\arguments{
\item{tables}{A list of tables (samples, phylogeny, compositions)}

\item{patient}{The patient to filter by}
}
\value{
A list of tables filtered by patient
}
\description{
Given a list of tables, filter them by patient.
}
