% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.clonevol.R
\name{jellyfisher.clonevol}
\alias{jellyfisher.clonevol}
\title{\if{html}{\out{
<Add Title>
}}}
\usage{
jellyfisher.clonevol(y, model = 1, explicit_parents = list())
}
\description{
\if{html}{\out{
<Add Description>
}}
}
