% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_patients}
\alias{select_patients}
\title{Filter Jellyfish tables by patient}
\usage{
select_patients(tables, patient)
}
\arguments{
\item{tables}{A list of tables (samples, phylogeny, compositions, ranks)}

\item{patient}{The patient or patients to filter by}
}
\value{
A list of tables filtered by patient
}
\description{
Given a list of tables, filter them by patient.
}
\examples{
jellyfisher_example_tables |>
  select_patients("EOC809")

}
