% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_parents}
\alias{set_parents}
\title{Set parents for samples}
\usage{
set_parents(tables, parents, unset_missing = FALSE)
}
\arguments{
\item{tables}{A list of tables (samples, phylogeny, compositions)}

\item{parents}{A named list of parents for each sample}

\item{unset_missing}{If TRUE, unset the parent for samples that are not in
the parent list}
}
\value{
A list of tables with parents set for each sample
}
\description{
Given a list of jellyfish input tables and a named list of parents for each
sample, set the parent for each sample.
}
\examples{
jellyfisher_example_tables |>
  filter_tables("EOC809") |>
  set_parents(list("EOC809_r1Bow1_DNA1" = "EOC809_p2Per1_cO_DNA2")) |>
  jellyfisher()

}
