% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_tables}
\alias{validate_tables}
\title{Validate Jellyfish tables}
\usage{
validate_tables(tables)
}
\arguments{
\item{tables}{A list of tables (samples, phylogeny, compositions, ranks)}
}
\description{
Superficially validate that the tables are in the correct format.
}
