% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonevol_functions.R
\name{extract_tables_from_clonevol}
\alias{extract_tables_from_clonevol}
\title{Extract samples, phylogeny, and subclonal compositions from ClonEvol results}
\usage{
extract_tables_from_clonevol(y, model = 1)
}
\arguments{
\item{y}{A ClonEvol object}

\item{model}{The model to extract. Defaults to 1}
}
\value{
A named list with three data frames: samples, phylogeny,
and compositions
}
\description{
Extracts data frames that can be used to create a Jellyfish plot.
}
\examples{
\dontrun{
# Run ClonEvol. Check the ClonEvol documentation for details.
y <- infer.clonal.models(...)
y <- convert.consensus.tree.clone.to.branch(y)

# Plot the results
extract_tables_from_clonevol(y, model = 1) |>
  jellyfisher()
}

}
