% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jellyfisher.R
\name{jellyfisher}
\alias{jellyfisher}
\title{Creates a Jellyfish plot}
\usage{
jellyfisher(
  samples,
  phylogeny,
  compositions,
  ranks = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{samples}{A data frame with samples}

\item{phylogeny}{A data frame with phylogeny}

\item{compositions}{A data frame with subclonal compositions}
}
\description{
Creates a Jellyfish plot from a samples, a phylogeny, and subclonal compositions.
The format of the data frames are described in Jellyfish documentation:
https://github.com/HautaniemiLab/jellyfish?tab=readme-ov-file#input-data
}
