% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jellyfisher.R
\name{jellyfisher}
\alias{jellyfisher}
\title{\if{html}{\out{
<Add Title>
}}}
\usage{
jellyfisher(
  samples,
  phylogeny,
  compositions,
  ranks = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
\if{html}{\out{
<Add Description>
}}
}
