% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonevol_functions.R
\name{jellyfisher_clonevol}
\alias{jellyfisher_clonevol}
\title{Create a Jellyfish plot from a ClonEvol object}
\usage{
jellyfisher_clonevol(y, model = 1, explicit_parents = list())
}
\arguments{
\item{y}{A ClonEvol object}

\item{model}{The model to extract from}
}
\description{
Takes a ClonEvol result object, extracts the necessary data, and creates
a Jellyfish plot.
}
